Sandbox systemd services
Restrict the proxy services' access to the host system.

This is the recomended list for Red Hat 7. More recent systemd
has a lot more options for this, but the Amazon Linux 2 base
we're using to experiment with nitro enclaves doesn't support them.

This is better than nothing, but the isolated-by-default model
of containers is easier to use properly.